% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additional_functions.R
\name{interpolate_theta}
\alias{interpolate_theta}
\title{Get \eqn{\theta(X)}}
\usage{
interpolate_theta(dat, X)
}
\arguments{
\item{dat}{data.frame which contains X and value of thate for that X}

\item{X}{vector for which new values of \eqn{\theta(X)} should be returned}
}
\value{
Returns \eqn{\theta(X)}
}
\description{
Get \eqn{\theta(X)}
}
