% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additional_functions.R
\name{k1}
\alias{k1}
\title{k1
\deqn{k_1=\left.\frac{1}{2v_0(y_{t-1})^2}\frac{\partial v_0(y_{t-1})^2}{\partial\theta(y_{t-1})}\right|_{\theta=\theta_0}}}
\usage{
k1(theta, x, func)
}
\arguments{
\item{theta}{A data.frame}

\item{x}{A number/vectors.}

\item{func}{scale function}
}
\value{
A value of score function.
}
\description{
k1
\deqn{k_1=\left.\frac{1}{2v_0(y_{t-1})^2}\frac{\partial v_0(y_{t-1})^2}{\partial\theta(y_{t-1})}\right|_{\theta=\theta_0}}
}
