% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additional_functions.R
\name{k2}
\alias{k2}
\title{k2}
\usage{
k2(theta, x, func)
}
\arguments{
\item{theta}{A data.frame}

\item{x}{A number/vectors.}

\item{func}{scale function}
}
\value{
A value of score function.
}
\description{
\deqn{k_2=\left.\frac{1}{v_0(y_{t-1})}\frac{\partial m_0(y_{t-1})^2}{\partial\theta(y_{t-1})}\right|_{\theta=\theta_0}}
}
