% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{pls}
\alias{pls}
\title{Pseudo Liklyhood Estimator for Location Scale model}
\usage{
pls(initial.theta, Y, X, func_s, func_m)
}
\arguments{
\item{initial.theta}{initial value of theta, A vector}

\item{Y}{parmeter of a function which is not to be optimized, usually \eqn{Y_t}}

\item{X}{regresor parameter can be X's or lag(Y_t)}

\item{func_s}{scale function}

\item{func_m}{location function}
}
\value{
Estimated location theta Robust
}
\description{
Description
}
