% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additional_functions.R
\name{qn_function}
\alias{qn_function}
\title{qn_function}
\usage{
qn_function(u, parameters = list())
}
\arguments{
\item{u}{A number}

\item{parameters}{A list with given parameters to function:
\code{k1m, k2m, A, cb, tau, func_mu, func_sigma, x, theta0}}
}
\value{
value of \eqn{q_n} function
}
\description{
\eqn{q_n}
}
