% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additional_functions.R
\name{qn_function_den}
\alias{qn_function_den}
\title{qn_function_den}
\usage{
qn_function_den(u, parameters = list())
}
\arguments{
\item{u}{A number}

\item{parameters}{A list with given parameters to function:
\code{k1m}, \code{k2m}, \code{A}, \code{cb}, \code{tau}, \code{func_mu},
\code{func_sigma}, \code{x}, \code{theta0}}
}
\value{
value of \eqn{q_n} function
}
\description{
\deqn{q^{den}_n(u):=\frac{c}{\|A(s(v;\theta_0)-\tau^{(0)})\|}}
Part of \eqn{\tau} calculation
}
