% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additional_functions.R
\name{qn_function_z}
\alias{qn_function_z}
\title{qn_function_z}
\usage{
qn_function_z(z, u, parameters)
}
\arguments{
\item{z}{A number}

\item{u}{A number}

\item{parameters}{A list with given parameters to function:
\code{k1m, k2m, A, cb, tau, func_mu, func_sigma, x, theta0}}
}
\value{
value of q_n function
}
\description{
\eqn{\underline{q_n}}
}
