% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{rls}
\alias{rls}
\title{Robust M-Estimator for Location Scale model}
\usage{
rls(theta, Y, X, c_bound, func_s, d_func_s, func_m, d_func_m, iterations = 5)
}
\arguments{
\item{theta}{initial value of theta, document later}

\item{Y}{parmeter of a function which is not to be optimized, usually \eqn{Y_t}}

\item{X}{regresor parameter can be X's or lag(Y_t)}

\item{c_bound}{bounding constant}

\item{func_s}{scale function}

\item{d_func_s}{derivative of scale function}

\item{func_m}{location function}

\item{d_func_m}{derivative of location function}

\item{iterations}{number of iterantions}
}
\value{
Estimated location theta Robust
}
\description{
Description
}
