% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additional_functions.R
\name{u_resids}
\alias{u_resids}
\title{Residuals}
\usage{
u_resids(y, x, theta, func_mu, func_sigma)
}
\arguments{
\item{y}{A number.}

\item{x}{A number/vectors.}

\item{theta}{A vector of lengths 2 of data.frame, depends on func_mu and
func_sigma.}

\item{func_mu}{location function}

\item{func_sigma}{scale function}
}
\value{
residual
}
\description{
Residuals
}
