% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additional_functions.R
\name{k2}
\alias{k2}
\title{k2}
\usage{
k2(theta, x, func)
}
\arguments{
\item{theta}{A data.frame}

\item{x}{A number/vectors.}

\item{func}{scale function}
}
\value{
A value of score function.
}
\description{
k2
}
