% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additional_functions.R
\name{qn_function}
\alias{qn_function}
\title{qn_function}
\usage{
qn_function(u, parameters = list())
}
\arguments{
\item{u}{A number}

\item{parameters}{A list with given parameters to function:
\code{k1m}, \code{k2m}, \code{A}, \code{cb}, \code{tau}, \code{func_mu},
\code{func_sigma}, \code{x}, \code{theta0}}
}
\value{
value of \eqn{q_n} function
}
\description{
\deqn{q_n(u):=(-k_{1}+k_{2}u+k_{1}u^2)\frac{c}{\|A(s(v;\theta_0)-\tau^{(0)})\|}}
}
