% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additional_functions.R
\name{qn_function_z}
\alias{qn_function_z}
\title{qn_function_z}
\usage{
qn_function_z(z, u, parameters)
}
\arguments{
\item{z}{A number}

\item{u}{A number}

\item{parameters}{A list with given parameters to function:
\code{k1m}, \code{k2m}, \code{A}, \code{cb}, \code{tau}, \code{func_mu},
\code{func_sigma}, \code{x}, \code{theta0}}
}
\value{
value of q_n function
}
\description{
\eqn{q_n(z):=q_n(u+z)\exp(-.5z^2)}
}
