% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{respls}
\alias{respls}
\title{Robust Kernel M-Estimator for Location Scale model}
\usage{
respls(theta, Y, X, c_bound, iterations = 5, bindwidths)
}
\arguments{
\item{theta}{initial value of theta, document later}

\item{Y}{parmeter of a function which is not to be optimized, usually \eqn{Y_t}}

\item{X}{regresor parameter can be X's or lag(Y_t)}

\item{c_bound}{bounding constant}

\item{iterations}{number of iterantions}

\item{bindwidths}{bindwidths should be used}
}
\value{
Estimated location theta Robust
}
\description{
Description
}
