% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additional_functions.R
\name{semi_est_func}
\alias{semi_est_func}
\title{Estimating function}
\usage{
semi_est_func(yt, thetas)
}
\arguments{
\item{yt}{A number.}

\item{thetas}{A vector of lengths 2.}
}
\value{
A value of score function.
}
\description{
Estimating function
}
